[YP-356] refactor: 작가 스튜디오 조회 훅 교체
