+++ src/components/creators/CreatorDetailView/components/CreatorIntroductionSection/components/CreatorIntroductionCategories.tsx
@@ -1,5 +1,6 @@
 import CategoryChip from "@/components/common/CategoryChip";
-import { useCreatorStudio } from "@/components/creators/CreatorDetailView/providers/CreatorStudioProvider";
+import useCreatorStudio from "@/components/creators/CreatorDetailView/hooks/useCreatorStudio";
+import useLoadingCreatorStudio from "@/components/creators/CreatorDetailView/hooks/useLoadingCreatorStudio";
 import { styled } from "@mui/material";
 
 const StyledCategoryList = styled("ul")`
@@ -10,9 +11,10 @@ const StyledCategoryList = styled("ul")`
 
 function CreatorIntroductionCategories() {
 	const studio = useCreatorStudio();
+	const isLoading = useLoadingCreatorStudio();
 
 	// TODO @김현규 스켈레톤 UI
-	if (!studio) {
+	if (isLoading) {
 		return <>로딩중</>;
 	}
