+++ src/components/creators/CreatorDetailView/components/CreatorIntroductionSection/components/CreatorIntroductionDescription.tsx
@@ -1,11 +1,13 @@
-import { useCreatorStudio } from "@/components/creators/CreatorDetailView/providers/CreatorStudioProvider";
+import useCreatorStudio from "@/components/creators/CreatorDetailView/hooks/useCreatorStudio";
+import useLoadingCreatorStudio from "@/components/creators/CreatorDetailView/hooks/useLoadingCreatorStudio";
 import { Box } from "@mui/material";
 
 function CreatorIntroductionDescription() {
 	const studio = useCreatorStudio();
+	const isLoading = useLoadingCreatorStudio();
 
 	// TODO @김현규 스켈레톤 UI
-	if (!studio) {
+	if (isLoading) {
 		return <>로딩중</>;
 	}
