+++ src/components/creators/CreatorDetailView/components/CreatorPortfoliosSection/index.tsx
@@ -1,5 +1,6 @@
 import InstagramMediaCard from "@/components/common/InstagramMediaCard";
-import { useCreatorStudio } from "@/components/creators/CreatorDetailView/providers/CreatorStudioProvider";
+import useCreatorStudio from "@/components/creators/CreatorDetailView/hooks/useCreatorStudio";
+import useLoadingCreatorStudio from "@/components/creators/CreatorDetailView/hooks/useLoadingCreatorStudio";
 import { styled } from "@mui/material";
 
 const StyledSection = styled("section")`
@@ -14,9 +15,10 @@ const StyledUl = styled("ul")`
 
 function CreatorPortfoliosSection() {
 	const studio = useCreatorStudio();
+	const isLoading = useLoadingCreatorStudio();
 
 	// TODO @김현규 스켈레톤 UI
-	if (!studio) {
+	if (isLoading) {
 		return <>로딩중</>;
 	}
