+++ src/components/creators/CreatorDetailView/index.ts
@@ -1,4 +1,3 @@
 import CreatorDetailView from "./CreatorDetailView";
 
 export default CreatorDetailView;
-export * from "./CreatorDetailView";
